3-prescription-reader/server/requirements.txt
# Based on virtual environment .venv-donut 
# Python 3.11.9
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
datasets==4.8.4
dill==0.4.1
evaluate==0.4.6
fastapi==0.135.3
filelock==3.25.2
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.10.2
idna==3.11
Jinja2==3.1.6
kiwisolver==1.5.0
lightning-utilities==0.15.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
networkx==3.6.1
numpy==2.4.4
onnx==1.21.0
onnx-ir==0.2.0
onnxscript==0.6.2
opencv-python-headless==4.13.0.92
packaging==26.0
pandas==3.0.2
pillow==12.2.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.0
pydantic_core==2.46.0
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
pytorch-lightning==2.6.1
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==15.0.0
safetensors==0.7.0
scipy==1.17.1
sentencepiece==0.2.1
shellingham==1.5.4
simsimd==6.5.16
six==1.17.0
starlette==1.0.0
stringzilla==4.6.0
sympy==1.14.0
tokenizers==0.22.2
torch==2.11.0
torchmetrics==1.9.0
torchvision==0.26.0
tqdm==4.67.3
transformers==5.5.4
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.44.0
watchfiles==1.1.1
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
